refactor: renamed and documented dart_indentifier_generator
